+++ pakyow-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `fix` **Load the environment before determining the server service count.**
+
+    *Related links:*
+    - [Pull Request #513][pr-513]
+
   * `chg` **Introduce `Pakyow.config.impolite` for controlling intro/goodbye logging.**
 
     *Related links:*
@@ -690,6 +695,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-513]: https://github.com/pakyow/pakyow/pull/513
 [pr-512]: https://github.com/pakyow/pakyow/pull/512
 [pr-511]: https://github.com/pakyow/pakyow/pull/511
 [pr-510]: https://github.com/pakyow/pakyow/pull/510